Split account storage into separate files.
With all of the account info centralized into one file, it was annoying to
work with individual accounts because API operations might lead to new tokens,
which required rewriting the whole file, which required each account to know
about the collection in which it was embedded. With a separate file per
account, they become fully orthogonal.