Track the association between files and accounts.
I'm pretty sure we're going to end up wanting this information. As in other
areas, things are a bit more complicated than you might think here, because
the same file might be accessible by more than one account known to the app.
So we need a side table.

It feels a little bit suboptimal to hold metadata about accounts in both a
JSON file and (a little bit) in the SQLite database, but <shrug>