Never mind, don't use People to get the account email address.
This failed for me in one case (my CfA account). I vaguely guess that this
organization, and presumably others, have permissions restrictions that
prevent the app from obtaining email addresses from the People API. But it
turns out that the Drive "About" API will return this information in a simpler
way! So let's do that.

The list of scopes that we request still seems reasonable, so I'm not changing
that.