Fix the "links" table.
The key is ... linkages between docs, which establish the folder graph, are
*also* account-specific. Two different accounts might both know about the same
file and folder, but in one account the file might be in the folder, but in
the other it might not. I think? Regardless, I'm pretty sure this is the best
way to approach the linkage graph.

Note that I'm editing the up.sql migration file, rather than creating a new
migration, since this chunk of work involved blowing away my whole database
anyway.