Add support for recent-listing shortcuts
This is a killer UX feature from my `bib` program. When a listing of documents
is printed, we give them short identifiers that can be used in subsequent
commands. This requires a new helper table but otherwise plugs in pretty
nicely with the common UX layer that I've been building up here.